feat(sc): add S3 logging Service Catalog product
Add a Service Catalog product to publish the S3 server access logging
bucket.

To be used, the product needs to be made part of a Service Catalog
portfolio. Because a portfolio can include many different products,
this repository should be included as a Git submodule in another repo
representing the overarching portfolio.

With the addition of new tests, the "default" test was also renamed to
"s3-log-bucket-with-default-parameters".

Associated issue: ShahradR/s3-logging#5